Api adjustments

+++ api-address-delivery/src/main/java/com/falzoni/api_address_delivery/services/AddressService.java
@@ -2,6 +2,8 @@ package com.falzoni.api_address_delivery.services;
 
 import com.falzoni.api_address_delivery.entities.Address;
 import com.falzoni.api_address_delivery.repositories.AddressRepository;
+import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,13 +19,20 @@ public class AddressService {
     }
 
     public Address find(int id) {
-        return this.repository.findById(id).orElse(null);
+        Address obj = this.repository.findById(id).orElse(null);
+        if(obj == null) {
+            throw new EntityNotFoundException("Registro não econtrado!");
+        }
+        return obj;
     }
 
     public void insert(Address obj) throws Exception {
         if(obj.getId() > 0) {
-            throw new Exception("Id identificado! Para atualização, deve-se utilizar o método PUT");
+            throw new IllegalArgumentException("Id identificado! Para atualização, deve-se utilizar o método PUT");
         }
+
+        this.validate(obj);
+
         this.repository.save(obj);
     }
 
@@ -31,10 +40,39 @@ public class AddressService {
         if(obj.getId() == 0) {
             throw new Exception("Id ausente! Para criação de um novo registro, deve-se utilizar o método POST");
         }
+
+        this.validate(obj);
+
         this.repository.save(obj);
     }
 
     public void delete(int id) {
+        if(!this.repository.existsById(id)) {
+            throw new EntityNotFoundException("Registro não econtrado!");
+        }
+
         this.repository.deleteById(id);
+    }
+
+    // private Methods
+    private void validate(Address obj) {
+        if(obj.getName() == null) {
+            throw new NullPointerException("O nome da rua do endereço é obrigatório");
+        }
+        if(obj.getNumber() <= 0) {
+            throw new NullPointerException("O número do endereço é obrigatório");
+        }
+        if(obj.getNeighborhood() == null) {
+            throw new NullPointerException("O bairro do endereço é obrigatório");
+        }
+        if(obj.getCity() == null) {
+            throw new NullPointerException("A cidade do endereço é obrigatória");
+        }
+        if(obj.getState() == null) {
+            throw new NullPointerException("O estado do endereço é obrigatório");
+        }
+        if(obj.getPostalCode() == null) {
+            throw new NullPointerException("O CEP do endereço é obrigatório");
+        }
     }
 }
